refactor(legacy-xls): split compatibility responsibilities
