+++ crates/converters/src/legacy_office/xls.rs
@@ -8,6 +8,19 @@ use into_markdown_core::{
     Provenance, ProvenanceKind, ValidationLimits,
 };
 
+mod binary;
+mod objects;
+mod preflight;
+mod wrapper;
+
+use binary::{read_u16, read_u32};
+use objects::retain_safe_images;
+use preflight::{
+    PreflightFlag, append_preflight_diagnostics, enforce_document_node_limit, preflight,
+    recover_continued_formula_string_caches,
+};
+use wrapper::{build_cfb_wrapper, cfb_wrapper_layout, normalize_raw_biff4};
+
 const WORKBOOK: &str = "Workbook";
 const BOF: u16 = 0x0809;
 const BOF4: u16 = 0x0409;
@@ -192,1589 +205,5 @@ pub(super) fn convert(
     Ok(output)
 }
 
-fn append_preflight_diagnostics(output: &mut ConverterOutput, preflight: &Preflight, part: &str) {
-    if preflight.has(PreflightFlag::ExternalBindings) {
-        output.diagnostics.push(Diagnostic {
-            code: "legacyOffice.xls.externalBindingsSkipped".into(),
-            severity: DiagnosticSeverity::Warning,
-            message: "external workbook bindings were retained as inert formula text and were not resolved"
-                .into(),
-            locator: Some(locator(part)),
-        });
-    }
-    if preflight.has(PreflightFlag::TailPadding) {
-        output.diagnostics.push(Diagnostic {
-            code: "legacyOffice.xls.trailingPaddingIgnored".into(),
-            severity: DiagnosticSeverity::Info,
-            message: "incomplete zero padding after a complete BIFF substream was ignored".into(),
-            locator: Some(locator(part)),
-        });
-    }
-    if matches!(preflight.biff_version, BIFF4 | BIFF5) {
-        output.diagnostics.push(Diagnostic {
-            code: "legacyOffice.xls.legacyBiffRecovered".into(),
-            severity: DiagnosticSeverity::Info,
-            message: format!(
-                "BIFF{} workbook data was converted through the bounded legacy reader",
-                if preflight.biff_version == BIFF4 { 4 } else { 5 }
-            ),
-            locator: Some(locator(part)),
-        });
-    }
-    if preflight.has(PreflightFlag::DimensionMetadata) {
-        output.diagnostics.push(Diagnostic {
-            code: "legacyOffice.xls.dimensionMetadataRecovered".into(),
-            severity: DiagnosticSeverity::Info,
-            message: "non-canonical Dimensions metadata was authenticated and omitted from the compatibility reader"
-                .into(),
-            locator: Some(locator(part)),
-        });
-    }
-    if preflight.has(PreflightFlag::FormulaCacheMetadata) {
-        output.diagnostics.push(Diagnostic {
-            code: "legacyOffice.xls.formulaStringCacheRecovered".into(),
-            severity: DiagnosticSeverity::Info,
-            message: "non-canonical reserved bytes in cached formula-string metadata were normalized without evaluating the formula"
-                .into(),
-            locator: Some(locator(part)),
-        });
-    }
-    if preflight.has(PreflightFlag::OptionalTailRecord) {
-        output.diagnostics.push(Diagnostic {
-            code: "legacyOffice.xls.optionalTailRecordIgnored".into(),
-            severity: DiagnosticSeverity::Info,
-            message: "a truncated trailing worksheet-view record after a complete BIFF substream was ignored"
-                .into(),
-            locator: Some(locator(part)),
-        });
-    }
-    if preflight.has(PreflightFlag::EmbeddedObjects) {
-        output.diagnostics.push(Diagnostic {
-            code: "legacyOffice.xls.embeddedObjectsSkipped".into(),
-            severity: DiagnosticSeverity::Warning,
-            message: "embedded OLE, drawing, and ActiveX objects were not executed or exported"
-                .into(),
-            locator: Some(locator(part)),
-        });
-    }
-}
-
-fn recover_continued_formula_string_caches(
-    workbook: &[u8],
-    output: &mut ConverterOutput,
-    part: &str,
-    budget: &mut LegacyBudget<'_>,
-    options: &ConversionOptions,
-) -> Result<usize, ConversionError> {
-    let mut cursor = 0_usize;
-    let mut sheet = None;
-    let mut next_sheet = 0_usize;
-    let mut recovered = 0_usize;
-    while cursor < workbook.len() {
-        budget.work(1, part)?;
-        let (kind, body, end) = biff_record(workbook, cursor, part)?;
-        if matches!(kind, BOF | BOF4) {
-            let substream = read_u16(body, 2, part)?;
-            if substream == 0x0010 {
-                sheet = Some(next_sheet);
-                next_sheet = next_sheet
-                    .checked_add(1)
-                    .ok_or_else(|| malformed(part, "BIFF worksheet count overflowed"))?;
-            }
-        } else if kind == EOF {
-            sheet = None;
-        } else if kind == FORMULA && has_noncanonical_formula_string_cache(body) && sheet.is_some()
-        {
-            let row = u32::from(read_u16(body, 0, part)?);
-            let column = u32::from(read_u16(body, 2, part)?);
-            if let Some(value) = decode_continued_formula_string(
-                workbook,
-                end,
-                part,
-                budget,
-                options.limits.max_field_bytes,
-            )? {
-                let replaced =
-                    replace_formula_cache(output, sheet.unwrap_or(0), row, column, &value);
-                if !replaced {
-                    cursor = end;
-                    continue;
-                }
-                recovered = recovered
-                    .checked_add(1)
-                    .ok_or_else(|| malformed(part, "formula-cache recovery count overflowed"))?;
-            }
-        }
-        cursor = end;
-    }
-    Ok(recovered)
-}
-
-fn biff_record<'a>(
-    bytes: &'a [u8],
-    cursor: usize,
-    part: &str,
-) -> Result<(u16, &'a [u8], usize), ConversionError> {
-    let header = bytes
-        .get(cursor..cursor.saturating_add(4))
-        .ok_or_else(|| malformed(part, "truncated BIFF record header"))?;
-    let kind = u16::from_le_bytes([header[0], header[1]]);
-    let length = usize::from(u16::from_le_bytes([header[2], header[3]]));
-    let body_start =
-        cursor.checked_add(4).ok_or_else(|| malformed(part, "BIFF record offset overflowed"))?;
-    let end = body_start
-        .checked_add(length)
-        .ok_or_else(|| malformed(part, "BIFF record length overflowed"))?;
-    let body =
-        bytes.get(body_start..end).ok_or_else(|| malformed(part, "truncated BIFF record body"))?;
-    Ok((kind, body, end))
-}
-
-fn decode_continued_formula_string(
-    workbook: &[u8],
-    cursor: usize,
-    part: &str,
-    budget: &mut LegacyBudget<'_>,
-    max_field_bytes: u64,
-) -> Result<Option<String>, ConversionError> {
-    let (mut kind, mut body, mut next) = biff_record(workbook, cursor, part)?;
-    if kind == SHARED_FORMULA {
-        (kind, body, next) = biff_record(workbook, next, part)?;
-    }
-    if kind != STRING {
-        return Err(malformed(part, "string-valued Formula is not followed by a String record"));
-    }
-    let characters = usize::from(read_u16(body, 0, part)?);
-    let flags = *body.get(2).ok_or_else(|| malformed(part, "truncated BIFF String flags"))?;
-    let mut offset = 3_usize;
-    if flags & 0x08 != 0 {
-        offset = offset
-            .checked_add(2)
-            .ok_or_else(|| malformed(part, "BIFF rich-string offset overflowed"))?;
-    }
-    if flags & 0x04 != 0 {
-        offset = offset
-            .checked_add(4)
-            .ok_or_else(|| malformed(part, "BIFF phonetic-string offset overflowed"))?;
-    }
-    if offset > body.len() {
-        return Err(malformed(part, "truncated BIFF String metadata"));
-    }
-    let initial_width = if flags & 1 == 0 { 1 } else { 2 };
-    let initial_capacity = body.len().saturating_sub(offset) / initial_width;
-    if characters <= initial_capacity {
-        return Ok(None);
-    }
-    if u64::try_from(characters).unwrap_or(u64::MAX) > max_field_bytes {
-        return Err(limit(
-            "max_field_bytes",
-            "continued cached formula string exceeds the field limit",
-        ));
-    }
-    let capacity = characters
-        .checked_mul(3)
-        .ok_or_else(|| limit("max_memory_bytes", "continued formula string size overflowed"))?;
-    let mut output = String::new();
-    output.try_reserve_exact(capacity).map_err(|error| {
-        limit("max_memory_bytes", format!("cannot reserve continued formula string: {error}"))
-    })?;
-    let mut remaining = characters;
-    let mut segment = &body[offset..];
-    let mut width = initial_width;
-    let mut pending_high_surrogate = None;
-    loop {
-        let available = segment.len() / width;
-        let take = remaining.min(available);
-        budget.work(u64::try_from(take).unwrap_or(u64::MAX), part)?;
-        if width == 1 {
-            if pending_high_surrogate.is_some() {
-                return Err(malformed(part, "BIFF String changes encoding within a surrogate"));
-            }
-            output.extend(segment[..take].iter().map(|byte| char::from(*byte)));
-        } else {
-            append_utf16_string_units(
-                &mut output,
-                &segment[..take * 2],
-                &mut pending_high_surrogate,
-                part,
-            )?;
-        }
-        remaining -= take;
-        if remaining == 0 {
-            break;
-        }
-        let (kind, continuation, end) = biff_record(workbook, next, part)?;
-        if kind != CONTINUE {
-            return Err(malformed(part, "BIFF String continuation is missing"));
-        }
-        let continuation_flags = *continuation
-            .first()
-            .ok_or_else(|| malformed(part, "empty BIFF String continuation"))?;
-        if continuation_flags & !1 != 0 {
-            return Err(malformed(part, "invalid BIFF String continuation flags"));
-        }
-        width = if continuation_flags == 0 { 1 } else { 2 };
-        segment = &continuation[1..];
-        next = end;
-    }
-    if pending_high_surrogate.is_some() {
-        return Err(malformed(part, "BIFF String ends with an incomplete surrogate"));
-    }
-    if u64::try_from(output.len()).unwrap_or(u64::MAX) > max_field_bytes {
-        return Err(limit(
-            "max_field_bytes",
-            "continued cached formula string exceeds the field limit",
-        ));
-    }
-    Ok(Some(output))
-}
-
-fn append_utf16_string_units(
-    output: &mut String,
-    bytes: &[u8],
-    pending_high_surrogate: &mut Option<u16>,
-    part: &str,
-) -> Result<(), ConversionError> {
-    for chunk in bytes.chunks_exact(2) {
-        let unit = u16::from_le_bytes([chunk[0], chunk[1]]);
-        if let Some(high) = pending_high_surrogate.take() {
-            if !(0xdc00..=0xdfff).contains(&unit) {
-                return Err(malformed(part, "invalid UTF-16 surrogate in BIFF String"));
-            }
-            let scalar = 0x1_0000 + ((u32::from(high) - 0xd800) << 10) + (u32::from(unit) - 0xdc00);
-            output.push(
-                char::from_u32(scalar)
-                    .ok_or_else(|| malformed(part, "invalid Unicode scalar in BIFF String"))?,
-            );
-        } else if (0xd800..=0xdbff).contains(&unit) {
-            *pending_high_surrogate = Some(unit);
-        } else if (0xdc00..=0xdfff).contains(&unit) {
-            return Err(malformed(part, "orphan UTF-16 surrogate in BIFF String"));
-        } else {
-            output.push(
-                char::from_u32(u32::from(unit))
-                    .ok_or_else(|| malformed(part, "invalid Unicode scalar in BIFF String"))?,
-            );
-        }
-    }
-    Ok(())
-}
-
-fn replace_formula_cache(
-    output: &mut ConverterOutput,
-    sheet_index: usize,
-    row: u32,
-    column: u32,
-    cache: &str,
-) -> bool {
-    let Some(blocks) = output
-        .document
-        .blocks
-        .iter_mut()
-        .filter_map(|node| match &mut node.block {
-            Block::Sheet { blocks, .. } => Some(blocks),
-            _ => None,
-        })
-        .nth(sheet_index)
-    else {
-        return false;
-    };
-    for node in blocks {
-        let Block::Table { rows, .. } = &mut node.block else { continue };
-        for table_row in rows {
-            for cell in &mut table_row.cells {
-                if !cell.blocks.iter().any(|node| {
-                    node.provenance
-                        .locator
-                        .cell
-                        .as_ref()
-                        .is_some_and(|reference| reference.row == row && reference.column == column)
-                }) {
-                    continue;
-                }
-                for node in &mut cell.blocks {
-                    let Block::Paragraph(inlines) = &mut node.block else { continue };
-                    for inline in inlines {
-                        match inline {
-                            Inline::Text { value, .. } => {
-                                cache.clone_into(value);
-                                return true;
-                            }
-                            Inline::Code(value) if value.starts_with('=') => {
-                                if let Some(marker) = value.rfind(" [cached: ") {
-                                    value.truncate(marker + " [cached: ".len());
-                                } else {
-                                    value.push_str(" [cached: ");
-                                }
-                                value.push_str(cache);
-                                value.push(']');
-                                return true;
-                            }
-                            _ => {}
-                        }
-                    }
-                }
-                return false;
-            }
-        }
-    }
-    false
-}
-
-fn enforce_document_node_limit(output: &ConverterOutput) -> Result<(), ConversionError> {
-    let Err(error) = output.document.validate() else { return Ok(()) };
-    if error.code == IrErrorCode::ResourceLimit
-        && output
-            .document
-            .validate_with_limits(&ValidationLimits {
-                max_nodes: usize::MAX,
-                ..ValidationLimits::default()
-            })
-            .is_ok()
-    {
-        return Err(limit(
-            "documentNodes",
-            format!(
-                "legacy XLS output exceeds the {} node limit",
-                into_markdown_core::MAX_DOCUMENT_NODES
-            ),
-        ));
-    }
-    Ok(())
-}
-
-#[derive(Clone, Copy)]
-#[repr(u8)]
-enum PreflightFlag {
-    ExternalBindings,
-    EmbeddedObjects,
-    TailPadding,
-    OptionalTailRecord,
-    DimensionMetadata,
-    FormulaCacheMetadata,
-}
-
-#[derive(Default)]
-struct PreflightFlags(u8);
-
-impl PreflightFlags {
-    fn insert(&mut self, flag: PreflightFlag) {
-        self.0 |= 1_u8 << flag as u8;
-    }
-
-    fn contains(&self, flag: PreflightFlag) -> bool {
-        self.0 & (1_u8 << flag as u8) != 0
-    }
-}
-
-#[derive(Default)]
-struct Preflight {
-    flags: PreflightFlags,
-    biff_version: u16,
-    logical_end: usize,
-}
-
-impl Preflight {
-    fn has(&self, flag: PreflightFlag) -> bool {
-        self.flags.contains(flag)
-    }
-}
-
-fn preflight(
-    bytes: &[u8],
-    part: &str,
-    budget: &mut LegacyBudget<'_>,
-    error_policy: ErrorPolicy,
-) -> Result<Preflight, ConversionError> {
-    let mut cursor = 0usize;
-    let mut biff_version = None;
-    let mut at_substream_boundary = false;
-    let mut zero_padding_start = None;
-    let mut result = Preflight { logical_end: bytes.len(), ..Preflight::default() };
-    while cursor < bytes.len() {
-        budget.work(1, part)?;
-        let Some(header) = bytes.get(cursor..cursor.saturating_add(4)) else {
-            if error_policy == ErrorPolicy::BestEffort
-                && (at_substream_boundary || zero_padding_start.is_some())
-                && bytes[cursor..].iter().all(|byte| *byte == 0)
-            {
-                result.flags.insert(PreflightFlag::TailPadding);
-                result.logical_end = zero_padding_start.unwrap_or(cursor);
-                break;
-            }
-            return Err(malformed(part, "truncated BIFF record header"));
-        };
-        let kind = u16::from_le_bytes([header[0], header[1]]);
-        let length = usize::from(u16::from_le_bytes([header[2], header[3]]));
-        let body_start = cursor
-            .checked_add(4)
-            .ok_or_else(|| malformed(part, "BIFF record offset overflowed"))?;
-        let end = body_start
-            .checked_add(length)
-            .ok_or_else(|| malformed(part, "BIFF record length overflowed"))?;
-        let Some(body) = bytes.get(body_start..end) else {
-            if error_policy == ErrorPolicy::BestEffort
-                && at_substream_boundary
-                && is_ignorable_tail_record(kind)
-            {
-                result.flags.insert(PreflightFlag::OptionalTailRecord);
-                result.logical_end = cursor;
-                break;
-            }
-            return Err(malformed(part, "truncated BIFF record body"));
-        };
-        if at_substream_boundary && kind == 0 && length == 0 {
-            zero_padding_start.get_or_insert(cursor);
-            cursor = end;
-            continue;
-        }
-        at_substream_boundary = false;
-        zero_padding_start = None;
-        match kind {
-            BOF | BOF4 => {
-                let version = if kind == BOF4 { BIFF4 } else { read_u16(body, 0, part)? };
-                let supported = version == BIFF8
-                    || (matches!(version, BIFF4 | BIFF5)
-                        && error_policy == ErrorPolicy::BestEffort);
-                if !supported {
-                    return Err(ConversionError::Unsupported {
-                        detail: format!(
-                            "XLS BIFF version 0x{version:04x} predates Office 97-2003 BIFF8"
-                        ),
-                    });
-                }
-                if biff_version.is_some_and(|current| current != version) {
-                    return Err(malformed(part, "BIFF substreams disagree on workbook version"));
-                }
-                biff_version = Some(version);
-            }
-            FILE_PASS => return Err(ConversionError::Encrypted),
-            EOF => at_substream_boundary = true,
-            DIMENSIONS => {
-                if preflight_dimensions(body, biff_version, part, budget, error_policy)? {
-                    result.flags.insert(PreflightFlag::DimensionMetadata);
-                }
-            }
-            FORMULA if has_noncanonical_formula_string_cache(body) => {
-                if error_policy == ErrorPolicy::Strict {
-                    return Err(malformed(part, "non-canonical cached formula-string metadata"));
-                }
-                result.flags.insert(PreflightFlag::FormulaCacheMetadata);
-            }
-            SUP_BOOK | EXTERN_SHEET => result.flags.insert(PreflightFlag::ExternalBindings),
-            OBJ | MSO_DRAWING_GROUP | MSO_DRAWING => {
-                result.flags.insert(PreflightFlag::EmbeddedObjects);
-            }
-            _ => {}
-        }
-        cursor = end;
-    }
-    if zero_padding_start.is_some() {
-        if error_policy == ErrorPolicy::Strict {
-            return Err(malformed(part, "zero padding follows the final BIFF substream"));
-        }
-        result.flags.insert(PreflightFlag::TailPadding);
-        result.logical_end = zero_padding_start.unwrap_or(bytes.len());
-    }
-    let Some(biff_version) = biff_version else {
-        return Err(malformed(part, "Workbook stream has no BIFF8 BOF record"));
-    };
-    result.biff_version = biff_version;
-    Ok(result)
-}
-
-fn is_ignorable_tail_record(kind: u16) -> bool {
-    matches!(kind, WINDOW1 | WINDOW2 | PANE | SELECTION | SCL)
-}
-
-fn has_noncanonical_formula_string_cache(body: &[u8]) -> bool {
-    body.get(6) == Some(&0)
-        && body.get(12..14) == Some(&[0xff, 0xff])
-        && body.get(7..12).is_some_and(|reserved| reserved.iter().any(|byte| *byte != 0))
-}
-
-fn preflight_dimensions(
-    body: &[u8],
-    biff_version: Option<u16>,
-    part: &str,
-    budget: &mut LegacyBudget<'_>,
-    error_policy: ErrorPolicy,
-) -> Result<bool, ConversionError> {
-    let legacy = matches!(biff_version, Some(BIFF4 | BIFF5));
-    let expected_length = if legacy { 10 } else { 14 };
-    let recovered_metadata = body.len() != expected_length;
-    if recovered_metadata && error_policy == ErrorPolicy::Strict {
-        return Err(malformed(part, "non-canonical BIFF Dimensions record length"));
-    }
-    let (first_row, last_row, first_column, last_column) = if legacy {
-        if body.len() < 8 {
-            return Err(malformed(part, "truncated BIFF5 Dimensions record"));
-        }
-        (
-            u64::from(read_u16(body, 0, part)?),
-            u64::from(read_u16(body, 2, part)?),
-            u64::from(read_u16(body, 4, part)?),
-            u64::from(read_u16(body, 6, part)?),
-        )
-    } else {
-        if body.len() < 12 {
-            return Err(malformed(part, "truncated BIFF8 Dimensions record"));
-        }
-        (
-            u64::from(read_u32(body, 0, part)?),
-            u64::from(read_u32(body, 4, part)?),
-            u64::from(read_u16(body, 8, part)?),
-            u64::from(read_u16(body, 10, part)?),
-        )
-    };
-    if last_row < first_row || last_column < first_column {
-        return Err(malformed(part, "BIFF8 Dimensions range is reversed"));
-    }
-    budget.table_shape(
-        usize::try_from(last_row - first_row).unwrap_or(usize::MAX),
-        usize::try_from(last_column - first_column).unwrap_or(usize::MAX),
-    )?;
-    Ok(recovered_metadata)
-}
-
-fn normalize_raw_biff4(bytes: &[u8]) -> Result<Vec<u8>, ConversionError> {
-    let first_length = usize::from(read_u16(bytes, 2, WORKBOOK)?);
-    let mut cursor = 4usize
-        .checked_add(first_length)
-        .ok_or_else(|| malformed(WORKBOOK, "raw BIFF4 BOF length overflowed"))?;
-    let global_prefix_start = cursor;
-    let mut global_prefix_end = None;
-    let mut formula_count = 0usize;
-    while cursor < bytes.len() {
-        let header_end = cursor
-            .checked_add(4)
-            .ok_or_else(|| malformed(WORKBOOK, "raw BIFF4 record offset overflowed"))?;
-        let header = bytes
-            .get(cursor..header_end)
-            .ok_or_else(|| malformed(WORKBOOK, "truncated raw BIFF4 record header"))?;
-        let kind = u16::from_le_bytes([header[0], header[1]]);
-        let length = usize::from(u16::from_le_bytes([header[2], header[3]]));
-        let end = cursor
-            .checked_add(4)
-            .and_then(|value| value.checked_add(length))
-            .ok_or_else(|| malformed(WORKBOOK, "raw BIFF4 record length overflowed"))?;
-        bytes.get(cursor..end).ok_or_else(|| malformed(WORKBOOK, "truncated raw BIFF4 record"))?;
-        if global_prefix_end.is_none() && kind == DIMENSIONS {
-            global_prefix_end = Some(cursor);
-        }
-        if global_prefix_end.is_none() && kind == EOF {
-            break;
-        }
-        if kind == 0x0406 {
-            formula_count = formula_count
-                .checked_add(1)
-                .ok_or_else(|| malformed(WORKBOOK, "raw BIFF4 Formula count overflowed"))?;
-        }
-        cursor = end;
-    }
-    let Some(global_prefix_end) = global_prefix_end else {
-        return Err(malformed(WORKBOOK, "raw BIFF4 stream has no Dimensions record"));
-    };
-
-    let global_prefix = bytes
-        .get(global_prefix_start..global_prefix_end)
-        .ok_or_else(|| malformed(WORKBOOK, "raw BIFF4 global record range is invalid"))?;
-    let formula_extra = formula_count
-        .checked_mul(4)
-        .ok_or_else(|| malformed(WORKBOOK, "normalized BIFF4 Formula bytes overflowed"))?;
-    let capacity = bytes
-        .len()
-        .checked_add(global_prefix.len())
-        .and_then(|value| value.checked_add(formula_extra))
-        .and_then(|value| value.checked_add(64))
-        .ok_or_else(|| malformed(WORKBOOK, "normalized BIFF4 size overflowed"))?;
-    let mut output = Vec::new();
-    output.try_reserve_exact(capacity).map_err(|error| {
-        limit("max_memory_bytes", format!("cannot reserve normalized BIFF4 stream: {error}"))
-    })?;
-    push_biff_record(&mut output, BOF, &[0x00, 0x04, 0x05, 0x00, 0, 0, 0, 0])?;
-    output.extend_from_slice(global_prefix);
-    append_normalized_biff4_sheet_header(&mut output)?;
-
-    cursor = 0;
-    while cursor < bytes.len() {
-        let body_start = cursor
-            .checked_add(4)
-            .ok_or_else(|| malformed(WORKBOOK, "raw BIFF4 record offset overflowed"))?;
-        let header = bytes
-            .get(cursor..body_start)
-            .ok_or_else(|| malformed(WORKBOOK, "truncated raw BIFF4 record header"))?;
-        let kind = u16::from_le_bytes([header[0], header[1]]);
-        let length = usize::from(u16::from_le_bytes([header[2], header[3]]));
-        let end = cursor
-            .checked_add(4)
-            .and_then(|value| value.checked_add(length))
-            .ok_or_else(|| malformed(WORKBOOK, "raw BIFF4 record length overflowed"))?;
-        let body = bytes
-            .get(body_start..end)
-            .ok_or_else(|| malformed(WORKBOOK, "truncated raw BIFF4 record"))?;
-        match kind {
-            BOF4 => {
-                let mut bof = body.to_vec();
-                let version = bof
-                    .get_mut(..2)
-                    .ok_or_else(|| malformed(WORKBOOK, "raw BIFF4 BOF body is truncated"))?;
-                version.copy_from_slice(&BIFF4.to_le_bytes());
-                push_biff_record(&mut output, BOF, &bof)?;
-            }
-            0x0406 => push_normalized_biff4_formula(&mut output, body)?,
-            _ => push_biff_record(&mut output, kind, body)?,
-        }
-        cursor = end;
-    }
-    Ok(output)
-}
-
-fn append_normalized_biff4_sheet_header(output: &mut Vec<u8>) -> Result<(), ConversionError> {
-    let sheet_name = b"Sheet 1";
-    let bound_sheet_record_bytes = 4usize
-        .checked_add(6)
-        .and_then(|value| value.checked_add(1 + sheet_name.len()))
-        .ok_or_else(|| malformed(WORKBOOK, "BIFF4 BoundSheet size overflowed"))?;
-    let sheet_offset = output
-        .len()
-        .checked_add(bound_sheet_record_bytes)
-        .and_then(|value| value.checked_add(4))
-        .ok_or_else(|| malformed(WORKBOOK, "BIFF4 sheet offset overflowed"))?;
-    let mut bound_sheet = Vec::with_capacity(bound_sheet_record_bytes - 4);
-    bound_sheet.extend_from_slice(&to_u32(sheet_offset)?.to_le_bytes());
-    bound_sheet.extend_from_slice(&[0, 0, u8::try_from(sheet_name.len()).unwrap_or(u8::MAX)]);
-    bound_sheet.extend_from_slice(sheet_name);
-    push_biff_record(output, 0x0085, &bound_sheet)?;
-    push_biff_record(output, EOF, &[])
-}
-
-fn push_normalized_biff4_formula(output: &mut Vec<u8>, body: &[u8]) -> Result<(), ConversionError> {
-    // BIFF4 places `cce` immediately after the two-byte option flags at offset 16.  The
-    // bounded downstream reader shares its BIFF8 record framing and expects the four-byte
-    // reserved field before `cce`.  Inserting that inert field preserves the cached value,
-    // options, token bytes, and BIFF4 token semantics selected by the BOF version.
-    if body.len() < 18 {
-        return Err(malformed(WORKBOOK, "raw BIFF4 Formula record is truncated"));
-    }
-    let normalized_length = body
-        .len()
-        .checked_add(4)
-        .ok_or_else(|| malformed(WORKBOOK, "normalized BIFF4 Formula size overflowed"))?;
-    output.extend_from_slice(&0x0006_u16.to_le_bytes());
-    output.extend_from_slice(&to_u16(normalized_length)?.to_le_bytes());
-    output.extend_from_slice(&body[..16]);
-    output.extend_from_slice(&[0; 4]);
-    output.extend_from_slice(&body[16..]);
-    Ok(())
-}
-
-fn push_biff_record(output: &mut Vec<u8>, kind: u16, body: &[u8]) -> Result<(), ConversionError> {
-    let length = to_u16(body.len())?;
-    output.extend_from_slice(&kind.to_le_bytes());
-    output.extend_from_slice(&length.to_le_bytes());
-    output.extend_from_slice(body);
-    Ok(())
-}
-
-#[derive(Clone, Copy)]
-struct CfbWrapperLayout {
-    stream_sectors: usize,
-    fat_sectors: usize,
-    difat_sectors: usize,
-    total_sectors: usize,
-    output_bytes: usize,
-}
-
-fn cfb_wrapper_layout(workbook_bytes: usize) -> Result<CfbWrapperLayout, ConversionError> {
-    let logical_bytes = workbook_bytes.max(4096);
-    let stream_sectors = logical_bytes.div_ceil(CFB_SECTOR_BYTES);
-    let mut fat_sectors = 1usize;
-    loop {
-        let difat_sectors =
-            fat_sectors.saturating_sub(CFB_HEADER_DIFAT_ENTRIES).div_ceil(CFB_DIFAT_ENTRIES);
-        let total_sectors = stream_sectors
-            .checked_add(1)
-            .and_then(|value| value.checked_add(fat_sectors))
-            .and_then(|value| value.checked_add(difat_sectors))
-            .ok_or_else(|| {
-                malformed(WORKBOOK, "CFB compatibility wrapper sector count overflowed")
-            })?;
-        let required_fat = total_sectors.div_ceil(CFB_FAT_ENTRIES);
-        if required_fat == fat_sectors {
-            let output_bytes = total_sectors
-                .checked_add(1)
-                .and_then(|value| value.checked_mul(CFB_SECTOR_BYTES))
-                .ok_or_else(|| malformed(WORKBOOK, "CFB compatibility wrapper size overflowed"))?;
-            return Ok(CfbWrapperLayout {
-                stream_sectors,
-                fat_sectors,
-                difat_sectors,
-                total_sectors,
-                output_bytes,
-            });
-        }
-        fat_sectors = required_fat;
-    }
-}
-
-fn build_cfb_wrapper(
-    workbook: &[u8],
-    recover_dimension_metadata: bool,
-    recover_formula_cache_metadata: bool,
-    biff_version: u16,
-    layout: CfbWrapperLayout,
-) -> Result<Vec<u8>, ConversionError> {
-    let mut output = Vec::new();
-    output.try_reserve_exact(layout.output_bytes).map_err(|error| {
-        limit("max_memory_bytes", format!("cannot reserve CFB compatibility wrapper: {error}"))
-    })?;
-    output.resize(layout.output_bytes, 0);
-    write_cfb_wrapper_header(&mut output, layout)?;
-    let fat_start = layout.stream_sectors + 1;
-    let difat_start = fat_start + layout.fat_sectors;
-    let stream_start = CFB_SECTOR_BYTES;
-    output[stream_start..stream_start + workbook.len()].copy_from_slice(workbook);
-    if recover_dimension_metadata {
-        patch_noncanonical_dimensions(&mut output, stream_start, workbook, biff_version)?;
-    }
-    if recover_formula_cache_metadata {
-        patch_noncanonical_formula_string_caches(&mut output, stream_start, workbook)?;
-    }
-
-    let directory_offset = sector_offset_in_wrapper(layout.stream_sectors)?;
-    write_directory_entry(
-        &mut output,
-        directory_offset,
-        DirectoryEntrySpec {
-            name: "Root Entry",
-            kind: 5,
-            left: CFB_FREE,
-            right: CFB_FREE,
-            child: 1,
-            start: CFB_END,
-            size: 0,
-        },
-    )?;
-    write_directory_entry(
-        &mut output,
-        directory_offset + 128,
-        DirectoryEntrySpec {
-            name: WORKBOOK,
-            kind: 2,
-            left: CFB_FREE,
-            right: CFB_FREE,
-            child: CFB_FREE,
-            start: 0,
-            size: u64::try_from(workbook.len().max(4096)).unwrap_or(u64::MAX),
-        },
-    )?;
-
-    let mut fat = Vec::new();
-    fat.try_reserve_exact(layout.total_sectors).map_err(|error| {
-        limit("max_memory_bytes", format!("cannot reserve CFB compatibility FAT: {error}"))
-    })?;
-    fat.resize(layout.total_sectors, CFB_FREE);
-    for (index, slot) in fat.iter_mut().take(layout.stream_sectors).enumerate() {
-        *slot = if index + 1 == layout.stream_sectors { CFB_END } else { to_u32(index + 1)? };
-    }
-    fat[layout.stream_sectors] = CFB_END;
-    for slot in fat.iter_mut().skip(fat_start).take(layout.fat_sectors) {
-        *slot = CFB_FAT;
-    }
-    for slot in fat.iter_mut().skip(difat_start).take(layout.difat_sectors) {
-        *slot = CFB_DIFAT;
-    }
-    write_cfb_wrapper_allocation_tables(&mut output, &fat, layout, fat_start, difat_start)?;
-    Ok(output)
-}
-
-fn patch_noncanonical_formula_string_caches(
-    output: &mut [u8],
-    stream_start: usize,
-    workbook: &[u8],
-) -> Result<(), ConversionError> {
-    let mut cursor = 0_usize;
-    while cursor < workbook.len() {
-        let header = workbook
-            .get(cursor..cursor.saturating_add(4))
-            .ok_or_else(|| malformed(WORKBOOK, "formula-cache patch found a truncated header"))?;
-        let kind = u16::from_le_bytes([header[0], header[1]]);
-        let length = usize::from(u16::from_le_bytes([header[2], header[3]]));
-        let end = cursor
-            .checked_add(4)
-            .and_then(|value| value.checked_add(length))
-            .ok_or_else(|| malformed(WORKBOOK, "formula-cache patch offset overflowed"))?;
-        let body = workbook
-            .get(cursor + 4..end)
-            .ok_or_else(|| malformed(WORKBOOK, "formula-cache patch found a truncated record"))?;
-        if kind == FORMULA && has_noncanonical_formula_string_cache(body) {
-            let reserved = stream_start
-                .checked_add(cursor)
-                .and_then(|value| value.checked_add(11))
-                .ok_or_else(|| malformed(WORKBOOK, "formula-cache patch offset overflowed"))?;
-            output
-                .get_mut(reserved..reserved + 5)
-                .ok_or_else(|| {
-                    malformed(WORKBOOK, "formula-cache patch is outside Workbook stream")
-                })?
-                .fill(0);
-        }
-        cursor = end;
-    }
-    Ok(())
-}
-
-fn patch_noncanonical_dimensions(
-    output: &mut [u8],
-    stream_start: usize,
-    workbook: &[u8],
-    biff_version: u16,
-) -> Result<(), ConversionError> {
-    let expected_length = if matches!(biff_version, BIFF4 | BIFF5) { 10 } else { 14 };
-    let mut cursor = 0_usize;
-    while cursor < workbook.len() {
-        let header = workbook
-            .get(cursor..cursor.saturating_add(4))
-            .ok_or_else(|| malformed(WORKBOOK, "compatibility patch found a truncated header"))?;
-        let kind = u16::from_le_bytes([header[0], header[1]]);
-        let length = usize::from(u16::from_le_bytes([header[2], header[3]]));
-        let end = cursor
-            .checked_add(4)
-            .and_then(|value| value.checked_add(length))
-            .ok_or_else(|| malformed(WORKBOOK, "compatibility patch offset overflowed"))?;
-        if end > workbook.len() {
-            return Err(malformed(WORKBOOK, "compatibility patch found a truncated record"));
-        }
-        if kind == DIMENSIONS && length != expected_length {
-            let physical = stream_start.checked_add(cursor).ok_or_else(|| {
-                malformed(WORKBOOK, "Dimensions compatibility patch offset overflowed")
-            })?;
-            output
-                .get_mut(physical..physical + 2)
-                .ok_or_else(|| {
-                    malformed(WORKBOOK, "Dimensions compatibility patch is outside Workbook stream")
-                })?
-                .copy_from_slice(&0xffff_u16.to_le_bytes());
-        }
-        cursor = end;
-    }
-    Ok(())
-}
-
-fn write_cfb_wrapper_header(
-    output: &mut [u8],
-    layout: CfbWrapperLayout,
-) -> Result<(), ConversionError> {
-    output[..8].copy_from_slice(b"\xd0\xcf\x11\xe0\xa1\xb1\x1a\xe1");
-    put_u16(output, 24, 0x003e)?;
-    put_u16(output, 26, 3)?;
-    put_u16(output, 28, 0xfffe)?;
-    put_u16(output, 30, 9)?;
-    put_u16(output, 32, 6)?;
-    put_u32(output, 44, to_u32(layout.fat_sectors)?)?;
-    let directory_sector = to_u32(layout.stream_sectors)?;
-    put_u32(output, 48, directory_sector)?;
-    put_u32(output, 56, 4096)?;
-    put_u32(output, 60, CFB_END)?;
-    let fat_start = layout.stream_sectors + 1;
-    let difat_start = fat_start + layout.fat_sectors;
-    put_u32(output, 68, if layout.difat_sectors == 0 { CFB_END } else { to_u32(difat_start)? })?;
-    put_u32(output, 72, to_u32(layout.difat_sectors)?)?;
-    for index in 0..CFB_HEADER_DIFAT_ENTRIES {
-        let value = if index < layout.fat_sectors.min(CFB_HEADER_DIFAT_ENTRIES) {
-            to_u32(fat_start + index)?
-        } else {
-            CFB_FREE
-        };
-        put_u32(output, 76 + index * 4, value)?;
-    }
-    Ok(())
-}
-
-fn write_cfb_wrapper_allocation_tables(
-    output: &mut [u8],
-    fat: &[u32],
-    layout: CfbWrapperLayout,
-    fat_start: usize,
-    difat_start: usize,
-) -> Result<(), ConversionError> {
-    for fat_index in 0..layout.fat_sectors {
-        let offset = sector_offset_in_wrapper(fat_start + fat_index)?;
-        for entry in 0..CFB_FAT_ENTRIES {
-            let value = fat.get(fat_index * CFB_FAT_ENTRIES + entry).copied().unwrap_or(CFB_FREE);
-            put_u32(output, offset + entry * 4, value)?;
-        }
-    }
-    let remaining_fat = layout.fat_sectors.saturating_sub(CFB_HEADER_DIFAT_ENTRIES);
-    for difat_index in 0..layout.difat_sectors {
-        let offset = sector_offset_in_wrapper(difat_start + difat_index)?;
-        for entry in 0..CFB_DIFAT_ENTRIES {
-            let index = difat_index * CFB_DIFAT_ENTRIES + entry;
-            let value = if index < remaining_fat {
-                to_u32(fat_start + CFB_HEADER_DIFAT_ENTRIES + index)?
-            } else {
-                CFB_FREE
-            };
-            put_u32(output, offset + entry * 4, value)?;
-        }
-        let next = if difat_index + 1 == layout.difat_sectors {
-            CFB_END
-        } else {
-            to_u32(difat_start + difat_index + 1)?
-        };
-        put_u32(output, offset + CFB_SECTOR_BYTES - 4, next)?;
-    }
-    Ok(())
-}
-
-#[derive(Clone, Copy)]
-struct DirectoryEntrySpec<'a> {
-    name: &'a str,
-    kind: u8,
-    left: u32,
-    right: u32,
-    child: u32,
-    start: u32,
-    size: u64,
-}
-
-fn write_directory_entry(
-    output: &mut [u8],
-    offset: usize,
-    entry: DirectoryEntrySpec<'_>,
-) -> Result<(), ConversionError> {
-    let mut encoded = entry.name.encode_utf16().collect::<Vec<_>>();
-    encoded.push(0);
-    if encoded.len() > 32 {
-        return Err(malformed(WORKBOOK, "CFB compatibility stream name is too long"));
-    }
-    for (index, unit) in encoded.iter().enumerate() {
-        put_u16(output, offset + index * 2, *unit)?;
-    }
-    put_u16(output, offset + 64, to_u16(encoded.len() * 2)?)?;
-    *output
-        .get_mut(offset + 66)
-        .ok_or_else(|| malformed(WORKBOOK, "CFB directory entry is truncated"))? = entry.kind;
-    *output
-        .get_mut(offset + 67)
-        .ok_or_else(|| malformed(WORKBOOK, "CFB directory entry is truncated"))? = 1;
-    put_u32(output, offset + 68, entry.left)?;
-    put_u32(output, offset + 72, entry.right)?;
-    put_u32(output, offset + 76, entry.child)?;
-    put_u32(output, offset + 116, entry.start)?;
-    put_u64(output, offset + 120, entry.size)
-}
-
-fn sector_offset_in_wrapper(sector: usize) -> Result<usize, ConversionError> {
-    sector
-        .checked_add(1)
-        .and_then(|value| value.checked_mul(CFB_SECTOR_BYTES))
-        .ok_or_else(|| malformed(WORKBOOK, "CFB compatibility sector offset overflowed"))
-}
-
-fn put_u16(output: &mut [u8], offset: usize, value: u16) -> Result<(), ConversionError> {
-    output
-        .get_mut(offset..offset + 2)
-        .ok_or_else(|| malformed(WORKBOOK, "CFB compatibility write is out of bounds"))?
-        .copy_from_slice(&value.to_le_bytes());
-    Ok(())
-}
-
-fn put_u32(output: &mut [u8], offset: usize, value: u32) -> Result<(), ConversionError> {
-    output
-        .get_mut(offset..offset + 4)
-        .ok_or_else(|| malformed(WORKBOOK, "CFB compatibility write is out of bounds"))?
-        .copy_from_slice(&value.to_le_bytes());
-    Ok(())
-}
-
-fn put_u64(output: &mut [u8], offset: usize, value: u64) -> Result<(), ConversionError> {
-    output
-        .get_mut(offset..offset + 8)
-        .ok_or_else(|| malformed(WORKBOOK, "CFB compatibility write is out of bounds"))?
-        .copy_from_slice(&value.to_le_bytes());
-    Ok(())
-}
-
-fn to_u16(value: usize) -> Result<u16, ConversionError> {
-    u16::try_from(value).map_err(|_| malformed(WORKBOOK, "CFB compatibility u16 overflowed"))
-}
-
-fn to_u32(value: usize) -> Result<u32, ConversionError> {
-    u32::try_from(value).map_err(|_| malformed(WORKBOOK, "CFB compatibility u32 overflowed"))
-}
-
-fn retain_safe_images(
-    bytes: &[u8],
-    part: &str,
-    output: &mut ConverterOutput,
-    budget: &mut LegacyBudget<'_>,
-) -> Result<(), ConversionError> {
-    let mut cursor = 0usize;
-    let mut count = 0usize;
-    while let Some((start, end, media_type)) = super::doc::find_image(&bytes[cursor..]) {
-        let start = cursor + start;
-        let end = cursor + end;
-        budget.raster(&bytes[start..end], media_type, "Workbook/image")?;
-        budget.asset(end - start, "Workbook/image")?;
-        count += 1;
-        let extension = if media_type == "image/png" { "png" } else { "jpg" };
-        let id = AssetId(format!("legacy-xls-asset-{count}"));
-        output.assets.push(Asset {
-            id: id.clone(),
-            filename: Some(format!("workbook-image-{count}.{extension}")),
-            media_type: media_type.into(),
-            bytes: bytes[start..end].to_vec(),
-            external_uri: None,
-        });
-        output.document.blocks.push(BlockNode {
-            id: NodeId(format!("legacy-xls-image-{count}")),
-            block: Block::Image { asset: id, alt: None },
-            provenance: Provenance {
-                kind: ProvenanceKind::NativeParser,
-                provider: PROVIDER_ID.into(),
-                locator: locator(part),
-                confidence: None,
-            },
-        });
-        cursor = end;
-    }
-    if count > 0 {
-        output.diagnostics.push(Diagnostic {
-            code: "legacyOffice.xls.imagePlacementRecovered".into(),
-            severity: DiagnosticSeverity::Warning,
-            message: "safe embedded image payloads were retained in workbook stream order because drawing anchors were incomplete"
-                .into(),
-            locator: Some(locator(part)),
-        });
-    }
-    Ok(())
-}
-
-fn read_u16(bytes: &[u8], offset: usize, part: &str) -> Result<u16, ConversionError> {
-    let raw = bytes.get(offset..offset + 2).ok_or_else(|| malformed(part, "truncated BIFF u16"))?;
-    Ok(u16::from_le_bytes([raw[0], raw[1]]))
-}
-
-fn read_u32(bytes: &[u8], offset: usize, part: &str) -> Result<u32, ConversionError> {
-    let raw = bytes.get(offset..offset + 4).ok_or_else(|| malformed(part, "truncated BIFF u32"))?;
-    Ok(u32::from_le_bytes([raw[0], raw[1], raw[2], raw[3]]))
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::msg::ole::CompoundFile;
-    use into_markdown_core::{Cell, ExecutionOptions, Inline, ResourceLimits};
-
-    fn budget<'a>(
-        options: &'a ConversionOptions,
-        context: &'a ExecutionContext,
-    ) -> LegacyBudget<'a> {
-        LegacyBudget::new(64, options, context).unwrap()
-    }
-
-    fn raw_biff4_with_label(text: &[u8]) -> Vec<u8> {
-        let mut bytes = Vec::new();
-        push_biff_record(&mut bytes, BOF4, &[0, 0, 0x10, 0, 0, 0]).unwrap();
-
-        let mut dimensions = Vec::new();
-        dimensions.extend_from_slice(&0u16.to_le_bytes());
-        dimensions.extend_from_slice(&1u16.to_le_bytes());
-        dimensions.extend_from_slice(&0u16.to_le_bytes());
-        dimensions.extend_from_slice(&1u16.to_le_bytes());
-        dimensions.extend_from_slice(&0u16.to_le_bytes());
-        push_biff_record(&mut bytes, DIMENSIONS, &dimensions).unwrap();
-
-        let mut label = vec![0; 6];
-        label.extend_from_slice(&u16::try_from(text.len()).unwrap().to_le_bytes());
-        label.extend_from_slice(text);
-        push_biff_record(&mut bytes, 0x0204, &label).unwrap();
-        push_biff_record(&mut bytes, EOF, &[]).unwrap();
-        bytes
-    }
-
-    fn convert_fixture(bytes: &[u8]) -> ConverterOutput {
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut compound_budget = LegacyBudget::new(bytes.len(), &options, &context).unwrap();
-        let compound = CompoundFile::open(bytes, &mut compound_budget).unwrap();
-        convert(bytes, compound.root(), &mut compound_budget, &options, &context, false).unwrap()
-    }
-
-    fn table(output: &ConverterOutput) -> (&str, &[into_markdown_core::TableRow]) {
-        let Block::Sheet { name, blocks } = &output.document.blocks[0].block else {
-            panic!("fixture did not emit a worksheet")
-        };
-        let Block::Table { rows, .. } = &blocks[0].block else {
-            panic!("fixture worksheet did not emit a table")
-        };
-        (name, rows)
-    }
-
-    fn cell_text(cell: &Cell) -> String {
-        let Some(block) = cell.blocks.first() else { return String::new() };
-        let Block::Paragraph(inlines) = &block.block else {
-            panic!("fixture cell did not emit a paragraph")
-        };
-        inlines
-            .iter()
-            .map(|inline| match inline {
-                Inline::Text { value, .. } | Inline::Code(value) => value.as_str(),
-                _ => panic!("fixture cell emitted an unexpected inline"),
-            })
-            .collect()
-    }
-
-    fn workbook_with_merge() -> Vec<u8> {
-        const FIXTURE: &[u8] =
-            include_bytes!("../../../../tools/macos-release/fixtures/normal.xls");
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut compound_budget = LegacyBudget::new(FIXTURE.len(), &options, &context).unwrap();
-        let compound = CompoundFile::open(FIXTURE, &mut compound_budget).unwrap();
-        let mut workbook = compound.root().stream(WORKBOOK).unwrap().to_vec();
-        let mut cursor = 0usize;
-        let mut final_eof = None;
-        while let Some(header) = workbook.get(cursor..cursor.saturating_add(4)) {
-            let kind = u16::from_le_bytes([header[0], header[1]]);
-            let length = usize::from(u16::from_le_bytes([header[2], header[3]]));
-            let Some(end) = cursor.checked_add(4).and_then(|value| value.checked_add(length))
-            else {
-                break;
-            };
-            if end > workbook.len() {
-                break;
-            }
-            if kind == EOF {
-                final_eof = Some(cursor);
-            }
-            cursor = end;
-        }
-        let mut merged = Vec::new();
-        push_biff_record(&mut merged, 0x00e5, &[1, 0, 0, 0, 0, 0, 0, 0, 1, 0]).unwrap();
-        workbook.splice(final_eof.unwrap()..final_eof.unwrap(), merged);
-        let layout = cfb_wrapper_layout(workbook.len()).unwrap();
-        build_cfb_wrapper(&workbook, false, false, BIFF8, layout).unwrap()
-    }
-
-    #[test]
-    fn rejects_pre_biff8_and_filepass() {
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut old = vec![0x09, 0x08, 4, 0, 0x00, 0x05, 0, 0];
-        assert!(matches!(
-            preflight(&old, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::Strict),
-            Err(ConversionError::Unsupported { .. })
-        ));
-        let recovered =
-            preflight(&old, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::BestEffort)
-                .unwrap();
-        assert_eq!(recovered.biff_version, BIFF5);
-        old[4..6].copy_from_slice(&BIFF8.to_le_bytes());
-        old.extend_from_slice(&[0x2f, 0, 0, 0]);
-        assert!(matches!(
-            preflight(&old, WORKBOOK, &mut budget(&options, &context), options.error_policy),
-            Err(ConversionError::Encrypted)
-        ));
-    }
-
-    #[test]
-    fn dimensions_use_table_resource_limits() {
-        let limits = ResourceLimits { max_table_rows: 10, ..ResourceLimits::default() };
-        let options = ConversionOptions { limits, ..ConversionOptions::default() };
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut bytes = vec![0x09, 0x08, 4, 0, 0, 6, 0, 0];
-        bytes.extend_from_slice(&[0x00, 0x02, 12, 0]);
-        bytes.extend_from_slice(&0u32.to_le_bytes());
-        bytes.extend_from_slice(&100u32.to_le_bytes());
-        bytes.extend_from_slice(&0u16.to_le_bytes());
-        bytes.extend_from_slice(&1u16.to_le_bytes());
-        assert!(matches!(
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), options.error_policy),
-            Err(ConversionError::ResourceLimit { limit: "max_table_rows", .. })
-        ));
-    }
-
-    #[test]
-    fn best_effort_accepts_only_zero_tail_after_complete_substream() {
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut bytes = vec![0x09, 0x08, 4, 0, 0, 6, 0, 0, 0x0a, 0, 0, 0, 0];
-        let recovered =
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::BestEffort)
-                .unwrap();
-        assert!(recovered.has(PreflightFlag::TailPadding));
-
-        assert!(matches!(
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::Strict,),
-            Err(ConversionError::Malformed { .. })
-        ));
-        *bytes.last_mut().unwrap() = 1;
-        assert!(matches!(
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::BestEffort,),
-            Err(ConversionError::Malformed { .. })
-        ));
-    }
-
-    #[test]
-    fn noncanonical_dimensions_are_omitted_only_in_best_effort() {
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut bytes = Vec::new();
-        push_biff_record(&mut bytes, BOF, &[0, 6, 0, 0]).unwrap();
-        let mut dimensions = Vec::new();
-        dimensions.extend_from_slice(&0u32.to_le_bytes());
-        dimensions.extend_from_slice(&1u32.to_le_bytes());
-        dimensions.extend_from_slice(&0u16.to_le_bytes());
-        dimensions.extend_from_slice(&1u16.to_le_bytes());
-        dimensions.extend_from_slice(&[0; 4]);
-        push_biff_record(&mut bytes, DIMENSIONS, &dimensions).unwrap();
-        push_biff_record(&mut bytes, EOF, &[]).unwrap();
-
-        let recovered =
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::BestEffort)
-                .unwrap();
-        assert!(recovered.has(PreflightFlag::DimensionMetadata));
-        assert!(matches!(
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::Strict,),
-            Err(ConversionError::Malformed { .. })
-        ));
-    }
-
-    #[test]
-    fn noncanonical_formula_string_cache_reserved_bytes_are_normalized_only_in_best_effort() {
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut bytes = Vec::new();
-        push_biff_record(&mut bytes, BOF, &[0, 6, 0, 0]).unwrap();
-        let formula_offset = bytes.len();
-        let mut formula = vec![0; 22];
-        formula[7] = 0x5a;
-        formula[12..14].fill(0xff);
-        push_biff_record(&mut bytes, FORMULA, &formula).unwrap();
-        push_biff_record(&mut bytes, EOF, &[]).unwrap();
-
-        let recovered =
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::BestEffort)
-                .unwrap();
-        assert!(recovered.has(PreflightFlag::FormulaCacheMetadata));
-        assert!(matches!(
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::Strict),
-            Err(ConversionError::Malformed { .. })
-        ));
-
-        let layout = cfb_wrapper_layout(bytes.len()).unwrap();
-        let wrapper = build_cfb_wrapper(&bytes, false, true, BIFF8, layout).unwrap();
-        let mut wrapper_budget = LegacyBudget::new(wrapper.len(), &options, &context).unwrap();
-        let compound = CompoundFile::open(&wrapper, &mut wrapper_budget).unwrap();
-        let workbook = compound.root().stream(WORKBOOK).unwrap();
-        assert_eq!(&workbook[formula_offset + 11..formula_offset + 16], &[0; 5]);
-    }
-
-    #[test]
-    fn compatibility_wrapper_is_a_bounded_readable_cfb() {
-        let mut workbook = raw_biff4_with_label(b"ok");
-        let dimensions_offset =
-            workbook.windows(4).position(|window| window == [0x00, 0x02, 0x0a, 0x00]).unwrap();
-        workbook[dimensions_offset + 2..dimensions_offset + 4]
-            .copy_from_slice(&12_u16.to_le_bytes());
-        workbook.splice(dimensions_offset + 14..dimensions_offset + 14, [0, 0]);
-        let layout = cfb_wrapper_layout(workbook.len()).unwrap();
-        let wrapper = build_cfb_wrapper(&workbook, true, false, BIFF4, layout).unwrap();
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut wrapper_budget = LegacyBudget::new(wrapper.len(), &options, &context).unwrap();
-        let compound = CompoundFile::open(&wrapper, &mut wrapper_budget).unwrap();
-        let recovered = compound.root().stream(WORKBOOK).unwrap();
-
-        assert_eq!(&recovered[dimensions_offset..dimensions_offset + 2], &[0xff, 0xff]);
-        assert_eq!(&recovered[..dimensions_offset], &workbook[..dimensions_offset]);
-    }
-
-    #[test]
-    fn large_compatibility_cfb_uses_precise_open_lease_boundary() {
-        let workbook = vec![0x5a; 16 * 1024 * 1024];
-        let layout = cfb_wrapper_layout(workbook.len()).unwrap();
-        let wrapper = build_cfb_wrapper(&workbook, false, false, BIFF8, layout).unwrap();
-
-        let options = ConversionOptions::default();
-        let measuring = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut measuring_budget = LegacyBudget::new(wrapper.len(), &options, &measuring).unwrap();
-        let compound = crate::msg::ole::CompoundFile::open_with_compatibility(
-            &wrapper,
-            &mut measuring_budget,
-            crate::msg::ole::CompoundCompatibility::LegacyOfficeBestEffort,
-        )
-        .unwrap();
-        let required = measuring.reserved_memory_bytes();
-        assert!(required > workbook.len() as u64);
-        assert!(required < wrapper.len() as u64 * 2);
-        drop(compound);
-        assert_eq!(measuring.reserved_memory_bytes(), 0);
-
-        let exact_options = ConversionOptions {
-            limits: ResourceLimits { max_memory_bytes: required, ..options.limits.clone() },
-            ..options.clone()
-        };
-        let exact =
-            ExecutionContext::new(ExecutionOptions::default(), exact_options.limits.clone());
-        let mut exact_budget = LegacyBudget::new(wrapper.len(), &exact_options, &exact).unwrap();
-        let compound = crate::msg::ole::CompoundFile::open_with_compatibility(
-            &wrapper,
-            &mut exact_budget,
-            crate::msg::ole::CompoundCompatibility::LegacyOfficeBestEffort,
-        )
-        .unwrap();
-        drop(compound);
-        assert_eq!(exact.reserved_memory_bytes(), 0);
-
-        let below_options = ConversionOptions {
-            limits: ResourceLimits { max_memory_bytes: required - 1, ..options.limits.clone() },
-            ..options
-        };
-        let below =
-            ExecutionContext::new(ExecutionOptions::default(), below_options.limits.clone());
-        let mut below_budget = LegacyBudget::new(wrapper.len(), &below_options, &below).unwrap();
-        let error = crate::msg::ole::CompoundFile::open_with_compatibility(
-            &wrapper,
-            &mut below_budget,
-            crate::msg::ole::CompoundCompatibility::LegacyOfficeBestEffort,
-        )
-        .unwrap_err();
-        assert!(matches!(error, ConversionError::ResourceLimit { .. }));
-        assert_eq!(below.reserved_memory_bytes(), 0);
-    }
-
-    #[test]
-    fn dense_noncanonical_dimensions_use_one_streaming_recovery_flag() {
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut bytes = Vec::new();
-        push_biff_record(&mut bytes, BOF, &[0, 6, 0, 0]).unwrap();
-        let mut dimensions = Vec::new();
-        dimensions.extend_from_slice(&0_u32.to_le_bytes());
-        dimensions.extend_from_slice(&1_u32.to_le_bytes());
-        dimensions.extend_from_slice(&0_u16.to_le_bytes());
-        dimensions.extend_from_slice(&1_u16.to_le_bytes());
-        dimensions.extend_from_slice(&[0; 4]);
-        for _ in 0..4_096 {
-            push_biff_record(&mut bytes, DIMENSIONS, &dimensions).unwrap();
-        }
-        push_biff_record(&mut bytes, EOF, &[]).unwrap();
-
-        let recovered =
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::BestEffort)
-                .unwrap();
-        assert!(recovered.has(PreflightFlag::DimensionMetadata));
-        assert_eq!(recovered.logical_end, bytes.len());
-        assert_eq!(context.reserved_memory_bytes(), 0);
-    }
-
-    #[test]
-    fn best_effort_ignores_only_truncated_view_metadata_after_complete_substream() {
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut bytes = Vec::new();
-        push_biff_record(&mut bytes, BOF, &[0, 6, 0, 0]).unwrap();
-        push_biff_record(&mut bytes, EOF, &[]).unwrap();
-        let complete_end = bytes.len();
-        bytes.extend_from_slice(&WINDOW1.to_le_bytes());
-        bytes.extend_from_slice(&10_u16.to_le_bytes());
-        bytes.extend_from_slice(&[1, 2]);
-
-        let recovered =
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::BestEffort)
-                .unwrap();
-        assert!(recovered.has(PreflightFlag::OptionalTailRecord));
-        assert_eq!(recovered.logical_end, complete_end);
-        assert!(
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::Strict)
-                .is_err()
-        );
-
-        bytes[complete_end..complete_end + 2].copy_from_slice(&DIMENSIONS.to_le_bytes());
-        assert!(
-            preflight(&bytes, WORKBOOK, &mut budget(&options, &context), ErrorPolicy::BestEffort)
-                .is_err()
-        );
-
-        let mut incomplete = Vec::new();
-        push_biff_record(&mut incomplete, BOF, &[0, 6, 0, 0]).unwrap();
-        incomplete.extend_from_slice(&WINDOW1.to_le_bytes());
-        incomplete.extend_from_slice(&10_u16.to_le_bytes());
-        incomplete.extend_from_slice(&[1, 2]);
-        assert!(
-            preflight(
-                &incomplete,
-                WORKBOOK,
-                &mut budget(&options, &context),
-                ErrorPolicy::BestEffort,
-            )
-            .is_err()
-        );
-    }
-
-    #[test]
-    fn xls_content_cell_order_display_values_and_merges_are_stable() {
-        const FIXTURE: &[u8] =
-            include_bytes!("../../../../tools/macos-release/fixtures/normal.xls");
-        let output = convert_fixture(FIXTURE);
-        let (name, rows) = table(&output);
-        assert_eq!(name, "Corpus");
-        let values = rows
-            .iter()
-            .map(|row| row.cells.iter().map(cell_text).collect::<Vec<_>>())
-            .collect::<Vec<_>>();
-        assert_eq!(
-            values,
-            [
-                ["Corpus", "=TRUE [cached: true]", "42.5"],
-                ["2024-01-01 00:00:00", "=SUM(1,2) [cached: 3]", "=cmd"],
-            ]
-        );
-        for (row_index, row) in rows.iter().enumerate() {
-            for (column_index, cell) in row.cells.iter().enumerate() {
-                let locator = &cell.blocks[0].provenance.locator;
-                assert_eq!(locator.sheet.as_deref(), Some("Corpus"));
-                let reference = locator.cell.as_ref().unwrap();
-                assert_eq!(reference.row, u32::try_from(row_index).unwrap());
-                assert_eq!(reference.column, u32::try_from(column_index).unwrap());
-            }
-        }
-
-        let merged = convert_fixture(&workbook_with_merge());
-        let (_, rows) = table(&merged);
-        assert_eq!(rows[0].cells.len(), 2);
-        assert_eq!(rows[0].cells[0].row_span, 1);
-        assert_eq!(rows[0].cells[0].column_span, 2);
-        assert_eq!(cell_text(&rows[0].cells[0]), "Corpus");
-        assert_eq!(cell_text(&rows[0].cells[1]), "42.5");
-    }
-
-    #[test]
-    fn biff4_formula_framing_preserves_value_options_and_tokens() {
-        let mut body = (0u8..16).collect::<Vec<_>>();
-        body.extend_from_slice(&3u16.to_le_bytes());
-        body.extend_from_slice(&[0x1e, 1, 0]);
-        let mut record = Vec::new();
-        push_normalized_biff4_formula(&mut record, &body).unwrap();
-
-        assert_eq!(&record[..2], &0x0006_u16.to_le_bytes());
-        let expected_record_size = u16::try_from(body.len())
-            .expect("fixture body fits BIFF record size")
-            .checked_add(4)
-            .expect("normalized formula record size fits u16");
-        assert_eq!(u16::from_le_bytes([record[2], record[3]]), expected_record_size);
-        assert_eq!(&record[4..20], &body[..16]);
-        assert_eq!(&record[20..24], &[0; 4]);
-        assert_eq!(&record[24..], &body[16..]);
-    }
-
-    #[test]
-    fn continued_formula_strings_are_bounded_and_decoded_without_evaluation() {
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut bytes = Vec::new();
-        push_biff_record(&mut bytes, STRING, &[5, 0, 0, b'a', b'b']).unwrap();
-        push_biff_record(&mut bytes, CONTINUE, &[0, b'c', b'd', b'e']).unwrap();
-        let mut conversion_budget = budget(&options, &context);
-        assert_eq!(
-            decode_continued_formula_string(
-                &bytes,
-                0,
-                WORKBOOK,
-                &mut conversion_budget,
-                options.limits.max_field_bytes,
-            )
-            .unwrap()
-            .as_deref(),
-            Some("abcde")
-        );
-
-        let mut limited_budget = budget(&options, &context);
-        assert!(matches!(
-            decode_continued_formula_string(&bytes, 0, WORKBOOK, &mut limited_budget, 4,),
-            Err(ConversionError::ResourceLimit { limit: "max_field_bytes", .. })
-        ));
-    }
-
-    #[test]
-    fn recovered_formula_cache_replaces_only_the_addressed_inert_cell() {
-        const FIXTURE: &[u8] =
-            include_bytes!("../../../../tools/macos-release/fixtures/normal.xls");
-        let mut output = convert_fixture(FIXTURE);
-        assert!(replace_formula_cache(&mut output, 0, 0, 1, "oracle text"));
-        assert!(replace_formula_cache(&mut output, 0, 0, 0, "cached-only text"));
-        let (_, rows) = table(&output);
-        assert_eq!(cell_text(&rows[0].cells[0]), "cached-only text");
-        assert_eq!(cell_text(&rows[0].cells[1]), "=TRUE [cached: oracle text]");
-        assert_eq!(cell_text(&rows[0].cells[2]), "42.5");
-    }
-
-    #[test]
-    fn raw_biff4_is_normalized_but_strict_mode_rejects_it() {
-        let bytes = raw_biff4_with_label(b"ok");
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut best_effort_budget = LegacyBudget::new(bytes.len(), &options, &context).unwrap();
-        let output = convert_raw(&bytes, &mut best_effort_budget, &options, &context).unwrap();
-
-        assert!(format!("{:?}", output.document.blocks).contains("ok"));
-        assert!(output.diagnostics.iter().any(|diagnostic| {
-            diagnostic.code == "legacyOffice.xls.rawBiffRecovered"
-                && diagnostic.severity == DiagnosticSeverity::Info
-        }));
-
-        let strict_options =
-            ConversionOptions { error_policy: ErrorPolicy::Strict, ..ConversionOptions::default() };
-        let strict_context =
-            ExecutionContext::new(ExecutionOptions::default(), strict_options.limits.clone());
-        let mut strict_budget =
-            LegacyBudget::new(bytes.len(), &strict_options, &strict_context).unwrap();
-        assert!(matches!(
-            convert_raw(&bytes, &mut strict_budget, &strict_options, &strict_context),
-            Err(ConversionError::Unsupported { .. })
-        ));
-    }
-
-    #[test]
-    fn raw_biff4_applies_recovery_state_and_emits_security_diagnostics() {
-        let mut bytes = raw_biff4_with_label(b"ok");
-        let dimensions =
-            bytes.windows(4).position(|window| window == [0x00, 0x02, 0x0a, 0x00]).unwrap();
-        bytes[dimensions + 2..dimensions + 4].copy_from_slice(&12_u16.to_le_bytes());
-        bytes.splice(dimensions + 14..dimensions + 14, [0, 0]);
-
-        let eof = bytes.windows(4).rposition(|window| window == [0x0a, 0x00, 0x00, 0x00]).unwrap();
-        let mut inert_metadata = Vec::new();
-        push_biff_record(&mut inert_metadata, SUP_BOOK, &[]).unwrap();
-        push_biff_record(&mut inert_metadata, OBJ, &[]).unwrap();
-        bytes.splice(eof..eof, inert_metadata);
-        bytes.extend_from_slice(&WINDOW1.to_le_bytes());
-        bytes.extend_from_slice(&10_u16.to_le_bytes());
-        bytes.extend_from_slice(&[1, 2]);
-
-        let options = ConversionOptions::default();
-        let context = ExecutionContext::new(ExecutionOptions::default(), options.limits.clone());
-        let mut conversion_budget = LegacyBudget::new(bytes.len(), &options, &context).unwrap();
-        let output = convert_raw(&bytes, &mut conversion_budget, &options, &context).unwrap();
-        for code in [
-            "legacyOffice.xls.dimensionMetadataRecovered",
-            "legacyOffice.xls.externalBindingsSkipped",
-            "legacyOffice.xls.embeddedObjectsSkipped",
-            "legacyOffice.xls.optionalTailRecordIgnored",
-        ] {
-            assert!(output.diagnostics.iter().any(|item| item.code == code), "missing {code}");
-        }
-        assert_eq!(
-            output
-                .diagnostics
-                .iter()
-                .filter(|item| item.code == "legacyOffice.xls.embeddedObjectsSkipped")
-                .count(),
-            1
-        );
-    }
-}
+mod tests;
