+++ tools/legacy_office_perf/__init__.py
@@ -1,0 +1,2 @@
+"""Legacy Office performance benchmark implementation."""
+
