+++ tools/legacy_office_perf/constants.py
@@ -1,0 +1,5 @@
+"""Shared benchmark constants."""
+
+FORMATS = ("doc", "ppt", "xls")
+MIB = 1024 * 1024
+
